+++ src/models/ObjectPickItem.ts
@@ -1,0 +1,6 @@
+import { QuickPickItem } from "vscode";
+
+export interface ObjectPickItem extends QuickPickItem {
+    id: string
+}
+